Update main.h
Significant inflation rate reduction added to base source code. Stake rate from 100%/year to 5%/year

It is up to the community to update their wallets, compile new binaries, etc. Then, if everyone (or the majority) switches to the new version, the supply will become more manageable

This is the first of a few actions that should hopefully motivate people to do things differently